config/requirements/common.txt
Django==3.1.3
djangorestframework==3.12.2
django-rest-auth
django-allauth
djangorestframework-jwt